% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_textcorpus_data.R
\name{add_description}
\alias{add_description}
\title{Add a \pkg{textcorpus} \code{description}}
\usage{
add_description(x, description.data = textcorpus::description, ...)
}
\arguments{
\item{x}{A description list that contains the required meta data 
about the corpus being added.  Use \code{\link{check_description}} to
check the template.}

\item{description.data}{The \pkg{textcorpus} description dataset.}

\item{\ldots}{ignored.}
}
\description{
Adds a \code{description} list to the \pkg{textcorpus} \code{description}.
}
\examples{
\dontrun{
desc_dat <- list(
    data = "mydat",
    genre = 'poem',
    subgenre = 'haiku',
    source = "https://url/to/nowhere.com",
    submitted_by = "Tyler Rinker",
    submitted_on = Sys.Date()
)

check_description(desc_dat)
add_description(desc_dat)
}
}
