% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_textcorpus_data.R
\name{check_description}
\alias{check_description}
\title{Check a \pkg{textcorpus} \code{description}}
\usage{
check_description(x, ...)
}
\arguments{
\item{x}{A description list that contains the required meta data 
about the corpus being added.  Use \code{\link{description_template}} to
start the template.}

\item{\ldots}{ignored.}
}
\description{
Ensures a \code{description} list is up \pkg{textcorpus} standards.
}
\examples{
desc_dat <- list(
    data = "mydat",
    genre = 'poem',
    subgenre = 'haiku',
    source = "https://url/to/nowhere.com",
    submitted_by = "Tyler Rinker",
    submitted_on = Sys.Date()
)

check_description(desc_dat)
}
