% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_textcorpus_data.R
\name{check_textcorpus_data}
\alias{check_textcorpus_data}
\alias{check_textcorpus_meta_data}
\title{Check if Data Meets \pkg{textcorpus} Standards}
\usage{
check_textcorpus_data(x, ...)

check_textcorpus_meta_data(x, meta, ...)
}
\arguments{
\item{x}{The main corpus data set.}

\item{meta}{The meta data set.}

\item{\ldots}{ignored.}
}
\description{
\code{check_textcorpus_data} - Ensure a corpus data set meets
\pkg{textcorpus} standards.  The data is expected to contain
\code{c('id', 'author', 'text')} columns.  The \code{"text"} column must
be a character vector.  Non-ASCII characters are not allowed in any columns.

\code{check_textcorpus_meta_data} - Ensure a corpus meta data set meets
\pkg{textcorpus} standards.  The data is expected to contain
an 'id' column at minimum.  Other `id` level meta data are accepted as well.
The `id` column must match the class of `x`'s `id` column.  Non-ASCII
characters are not allowed in any columns.
}
\examples{
library(dplyr)
library(stringi)

corp <- data_frame(
    id = stri_rand_strings(10, 10),
    author = c('sam', 'cal', 'sue', 'bob', 'sal', 'pam', 'pat', 'joe', 'arr', 'nmr'),
    text = stri_rand_lipsum(10),
    state = state.name[1:10],
    month = month.name[1:10]
)

check_textcorpus_data(corp)
check_textcorpus_meta_data(corp, corp[c('id', 'state', 'month')])
}

