% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textcorpus-package.R
\docType{data}
\name{description}
\alias{description}
\title{Descriptions of \pkg{textcorpus} Datasets}
\format{A tibble with 6 variables}
\usage{
data(description)
}
\description{
A searchable meta dataset on the available datasets in \pkg{textcorpus}.
}
\details{
\itemize{
  \item id. An id the maps between the \code{corpus} and \code{meta} datasets
  \item author. The speaker/writer of the text element
  \item text. The text variable
  \item order. VARIABLE DESCRIPTION
}

\itemize{
  \item data. The name of available corpus data set
  \item genre. The genere of the data set (e.g., poem, literature, song, debate, transcript, etc.)
  \item subgenre. A subgenre of the data set (e.g., political, educational, folk, country, comedy, etc.)
  \item source. Where the data was obtained from
  \item submitted_by. Who submitted the data to \pkg{textcorpus}
  \item submitted_on. The date when the data was submitted
}
}
