% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_textcorpus_data.R
\name{description_template}
\alias{description_template}
\title{Title}
\usage{
description_template(copy2clip = TRUE)
}
\arguments{
\item{copy2clip}{logical.  If \code{TRUE} the template is copied to the
clipboard.}
}
\description{
Description
}
\examples{
description_template(FALSE)
}
