% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_textcorpus_data.R
\name{prepare_textcorpus_data}
\alias{prepare_textcorpus_data}
\title{Perpare a Corpus Data Set for Inclusion in the \pkg{textcorpus} Package}
\usage{
prepare_textcorpus_data(x, corpus.cols = c("id", "author", "text"),
  meta = NULL, ...)
}
\arguments{
\item{x}{The main corpus data set.}

\item{corpus.cols}{The corpus columns found in \code{x}.  At the very least
\code{c('id', 'author', 'text')} are required columns.  The user may supply
the entire data set for \code{x} and specify meta data columns to peel off
and make distinct or may pass a reduced number of columns to \code{x} and
pass the matching meta data through the \code{meta} argument.}

\item{meta}{The optional meta data set.}

\item{\ldots}{ignored.}
}
\value{
Returns a list of a main corpus text data and meta data with a
matching `id` key.
}
\description{
Prepare a corpus data.frame for inclusion in the \pkg{textcorpus} package.
}
\examples{
library(dplyr)
library(stringi)

corp <- data_frame(
    id = stri_rand_strings(10, 10),
    author = c('sam', 'cal', 'sue', 'bob', 'sal', 'pam', 'pat', 'joe', 'arr', 'nmr'),
    text = stri_rand_lipsum(10),
    state = state.name[1:10],
    month = month.name[1:10]
)

prepare_textcorpus_data(corp)
prepare_textcorpus_data(corp, corpus.cols = c('id', 'author', 'text', 'state'))
}

